Extract useWatchedTime hook from Home
